refactor attempt 1
